old-projects/goards-api/goards_api/goards_api/wsgi.py
"""
WSGI config for goards_api project.

It exposes the WSGI callable as a module-level variable named ``application``.

For more information on this file, see
https://docs.djangoproject.com/en/1.10/howto/deployment/wsgi/
"""

import os

import dotenv
from django.core.wsgi import get_wsgi_application

dotenv.read_dotenv(
    os.path.join(
        os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__)))),
        'goards_api.env'
    )
)
os.environ.setdefault("DJANGO_SETTINGS_MODULE", "goards_api.settings")

application = get_wsgi_application()
