old-projects/goards-api/requirements/base.txt
appdirs==1.4.2
appnope==0.1.0
coreapi==2.3.0
coreschema==0.0.4
coverage==4.0.3
decorator==4.0.11
Django==1.11.1
django-cors-headers==2.0.2
django-dotenv==1.4.1
djangorestframework==3.6.2
djangorestframework-jwt==1.9.0
ipython==5.3.0
ipython-genutils==0.1.0
itypes==1.1.0
Jinja2==2.9.5
Markdown==2.6.8
MarkupSafe==1.0
packaging==16.8
pbr==2.0.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.13
ptyprocess==0.5.1
Pygments==2.2.0
PyJWT==1.4.2
pyparsing==2.1.10
python-coveralls==2.9.1
python-dateutil==2.6.0
pytz==2017.2
PyYAML==3.12
requests==2.13.0
simplegeneric==0.8.1
six==1.10.0
stevedore==1.21.0
traitlets==4.3.2
uritemplate==3.0.0
virtualenv==15.1.0
virtualenv-clone==0.2.6
wcwidth==0.1.7
